backend/core/security.py
# backend/core/security.py
from datetime import datetime, timedelta
from typing import Optional
import hashlib
import hmac
import os

from jose import JWTError, jwt
from core.config import settings


def hash_password(plain: str) -> str:
    """Hash password using PBKDF2-SHA256 — no bcrypt needed."""
    salt = os.urandom(32)
    key = hashlib.pbkdf2_hmac("sha256", plain.encode(), salt, 310000)
    return salt.hex() + ":" + key.hex()


def verify_password(plain: str, hashed: str) -> bool:
    """Verify a PBKDF2-SHA256 hashed password."""
    try:
        salt_hex, key_hex = hashed.split(":")
        salt = bytes.fromhex(salt_hex)
        key = bytes.fromhex(key_hex)
        new_key = hashlib.pbkdf2_hmac("sha256", plain.encode(), salt, 310000)
        return hmac.compare_digest(key, new_key)
    except Exception:
        return False


def create_access_token(subject: str, expires_delta: Optional[timedelta] = None) -> str:
    expire = datetime.utcnow() + (
        expires_delta or timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    )
    payload = {"sub": str(subject), "exp": expire, "iat": datetime.utcnow()}
    return jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.ALGORITHM)


def decode_access_token(token: str) -> Optional[str]:
    try:
        data = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
        return data.get("sub")
    except JWTError:
        return None