requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.4
hiredis==2.3.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# AI / ML
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# Claude API
anthropic==0.26.1

# HTTP Client
httpx==0.27.0

# Env
python-dotenv==1.0.1

# WebSockets
websockets==12.0
